mds/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
freeze==3.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
